ランキング機能の作成

+++ app/Http/Controllers/TodosController.php
@@ -6,23 +6,44 @@ use Illuminate\Http\Request;
 use App\Todo;
 use App\Time;
 use App\Http\Requests\TodoRequest;
+use App\User;
+use Auth;
 
 class TodosController extends Controller
 {
-    public function index(Todo $todos)
+    public function index(Todo $todo)
     {
-        $todos = Todo::all();
+        // $todo = Todo::where('user_id', Auth::id());
+        // dd($todo);
+        $todo = Todo::all();
         
-        
-        return view('tasks/index')->with('todos',$todos);
+        return view('tasks/index')->with('todos',$todo);
     }
     
     public function elapsedTime(Todo $todos)
     {
-        $todos = Todo::all();
+        $todos = Todo::where('user_id', Auth::id())->get();
+        // dd($todos);
+        // $todos = Todo::all();
+        // dd($todos);
         return view('tasks/index')->with(['todos' => $todos]);
         // dd($elapsed_time);
         
+    }
+    
+    public function endding(Todo $todo)
+    {
+        $todo = Todo::all();
+        
+         // userごとの本日のトータルを出す
+        $todos = Todo::where('user_id', Auth::id())->get();
+        $total = 0;
+        for($i = 0; $i < count($todos); $i++) {
+            $total = $total + $todos[$i]->tasks_time;
+        }
+        dd($total);
+        
+        return view('times/end')->with(['todo' => $todo, 'total' => $total]);
     }
     
     
@@ -41,37 +62,30 @@ class TodosController extends Controller
     public function store(Todo $todo, TodoRequest $request)
     {
         $input = $request['todo'];
+        $input += ['user_id' => $request->user()->id];
         $todo->fill($input)->save();
         return redirect('/tasks')->with(['todo' => $todo]);
     }
     
     public function destroy(Todo $todo)
     {
-        $todo->delete();
-        return redirect('/tasks'->with(['todo' => $todo]));
+        if($todo->achievement_task == 0){
+            $todo->delete();
+        }
+        return redirect('/tasks')->with(['todo' => $todo]);
     }
     
-    public function achieve(Todo $todo)
+    public function achieve(Todo $todo, User $user)
     {
-        // $todo = Todo::find({{ $todo->id }});
-        // $todo = Todo::all();
-        // dd($todo[0]);
-        // $b = $todo->id;
-        // dd($b);
-        // $todos = Todo::where('id', 3)->first();
-        // dd($todos->id);
-        // $a = $todo->id;
-        // dd($a);
-    
+        $todos = Todo::where('id', $todo->id)->first();
         $todo->fill(['achievement_task' => 1])->save();
-        // dd($todos);
-        // $achieve = Todo::where('achievement_task', true)->();
-        // dd($achieve);
-        // $a = $todo->achievement_task;
-        // // dd($a);
-        // $input = 1;
-        // $todo->fill($input)->save();
-        return redirect('/tasks');
+        
+        $user = User::where('id', Auth::id())->first();
+        
+        $task_count = $user->total_task + 1;
+        $user->fill(['total_task' => $task_count])->save();
+        
+        return redirect('/tasks')->with(['todo' => $todo]);
     }
